@@ -7,7 +7,7 @@ Python CAMB interface (https://camb.info)
 __author__ = "Antony Lewis"
 __contact__ = "antony at cosmologist dot info"
 __url__ = "https://camb.readthedocs.io"
-__version__ = "1.0.11.1"
+__version__ = "1.1.2.1"
 
 from . import baseconfig
 
@@ -15,7 +15,7 @@ baseconfig.check_fortran_version(__version__)
 from .baseconfig import CAMBFortranError, CAMBError, CAMBValueError, CAMBUnknownArgumentError, CAMBParamRangeError
 from .camb import get_results, get_transfer_functions, get_background, \
     get_age, get_zre_from_tau, set_feedback_level, set_params, get_matter_power_interpolator, \
-    set_params_cosmomc, read_ini, run_ini
+    set_params_cosmomc, read_ini, run_ini, get_valid_numerical_params
 from . import model
 from . import initialpower
 from . import reionization
