@@ -1,20 +1,14 @@
 from .baseconfig import camblib, CAMBError, CAMBValueError, CAMBUnknownArgumentError, np
-from ctypes import c_float, c_int, c_double, c_bool, POINTER, byref
+from ctypes import c_double, c_bool, POINTER, byref
 import ctypes
 from . import model, constants
 from ._config import config
 from .model import CAMBparams
 from .results import CAMBdata, MatterTransferData, ClTransferData
-
 import logging
-import six
 import os
-
-if six.PY3:
-    from inspect import getfullargspec as getargspec
-else:
-    # noinspection PyDeprecation
-    from inspect import getargspec
+import numbers
+from inspect import getfullargspec
 
 _debug_params = False
 
@@ -42,17 +36,20 @@ def get_results(params):
     return res
 
 
-def get_transfer_functions(params):
+def get_transfer_functions(params, only_time_sources=False):
     """
     Calculate transfer functions for specified parameters and return :class:`~.results.CAMBdata` instance for
     getting results and subsequently calculating power spectra.
 
     :param params: :class:`.model.CAMBparams` instance
+    :param only_time_sources: does not calculate the CMB l,k transfer functions and does not apply any non-linear
+                              correction scaling. Results with only_time_sources=True can therefore be used with
+                              different initial power spectra to get consistent non-linear lensed spectra.
     :return: :class:`~.results.CAMBdata` instance
     """
 
     res = CAMBdata()
-    res.calc_transfers(params)
+    res.calc_transfers(params, only_transfers=True, only_time_sources=only_time_sources)
     return res
 
 
@@ -145,8 +142,8 @@ def set_params(cp=None, verbose=False, **params):
     used_params = set()
 
     def do_set(setter):
-        kwargs = {k: params[k] for k in getargspec(setter).args[1:] if k in params}
-        used_params.update(kwargs.keys())
+        kwargs = {kk: params[kk] for kk in getfullargspec(setter).args[1:] if kk in params}
+        used_params.update(kwargs)
         if kwargs:
             if verbose:
                 logging.warning('Calling %s(**%s)' % (setter.__name__, kwargs))
@@ -182,6 +179,36 @@ def set_params(cp=None, verbose=False, **params):
             else:
                 raise CAMBUnknownArgumentError("Unrecognized parameter: %s" % k)
     return cp
+
+
+def get_valid_numerical_params(transfer_only=False, **class_names):
+    """
+    Get numerical parameter names that are valid input to :func:`set_params`
+
+    :param transfer_only: if True, exclude parameters that affect only initial power spectrum or non-linear model
+    :param class_names: class name parameters that will be used by :meth:`.model.CAMBparams.set_classes`
+    :return: set of valid input parameter names for :func:`set_params`
+    """
+    cp = CAMBparams()
+    cp.set_classes(**class_names)
+    params = set()
+
+    def extract_params(set_func):
+        pars = getfullargspec(set_func)
+        for arg, v in zip(pars.args[1:], pars.defaults[1:]):
+            if (isinstance(v, numbers.Number) or v is None) and 'version' not in arg:
+                params.add(arg)
+
+    extract_params(cp.DarkEnergy.set_params)
+    extract_params(cp.set_cosmology)
+    if not transfer_only:
+        extract_params(cp.InitPower.set_params)
+        extract_params(cp.NonLinearModel.set_params)
+    for f, tp in cp._fields_:
+        if not f.startswith('_') and tp == ctypes.c_double:
+            params.add(f)
+    return params - {'max_eta_k_tensor', 'max_eta_k', 'neutrino_hierarchy', 'standard_neutrino_neff',
+                     'pivot_scalar', 'num_massive_neutrinos', 'num_nu_massless'}
 
 
 def set_params_cosmomc(p, num_massive_neutrinos=1, neutrino_hierarchy='degenerate', halofit_version='mead',
@@ -247,7 +274,7 @@ def run_ini(ini_filename, no_validate=False):
     run_inifile = camblib.__camb_MOD_camb_runinifile
     run_inifile.argtypes = [ctypes.c_char_p, POINTER(ctypes.c_long)]
     run_inifile.restype = c_bool
-    s = ctypes.create_string_buffer(six.b(ini_filename))
+    s = ctypes.create_string_buffer(ini_filename.encode("latin-1"))
     if not run_inifile(s, ctypes.c_long(len(ini_filename))):
         config.check_global_error('run_ini')
 
@@ -268,7 +295,7 @@ def read_ini(ini_filename, no_validate=False):
     read_inifile = camblib.__camb_MOD_camb_readparamfile
     read_inifile.argtypes = [POINTER(CAMBparams), ctypes.c_char_p, POINTER(ctypes.c_long)]
     read_inifile.restype = ctypes.c_bool
-    s = ctypes.create_string_buffer(six.b(ini_filename))
+    s = ctypes.create_string_buffer(ini_filename.encode("latin-1"))
     if not read_inifile(cp, s, ctypes.c_long(len(ini_filename))):
         config.check_global_error('read_ini')
     return cp
